py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "atlas-init"
dynamic = ["version"]
description = ''
readme = "../readme.md"
requires-python = ">=3.12"
license = "MIT"
keywords = []
authors = [{ name = "EspenAlbert", email = "albertespen@gmail.com" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "boto3==1.34.55",
  "GitPython==3.1.42",
  "model-lib==0.0.30",
  "mypy-boto3-cloudformation==1.34.66",
  "orjson==3.9.15",
  "pydantic-settings==2.2.1",
  "requests==2.31.0",
  "rich==13.7.1",
  "stringcase==1.2.0",
  "tenacity==8.2.3",
  "typer==0.12.0",
]

[project.scripts]
atlasci = "atlas_init.cli:typer_main"

[project.urls]
Documentation = "https://github.com/EspenAlbert/atlas-init#readme"
Issues = "https://github.com/EspenAlbert/atlas-init/issues"
Source = "https://github.com/EspenAlbert/atlas-init"

[tool.hatch.version]
path = "atlas_init/__init__.py"

[tool.hatch.envs.default]
installer = "uv"

[tool.hatch.envs.hatch-test]
installer = "uv"
default-args = ["test_atlas_init", "--capture=no", "--tb=short", "-vv"]
extra-dependencies = [
  "hatchling==1.24.2",
  "pytest-cov==5.0.0"
] # used to make it simpler to sync venv with IDE

[tool.coverage.run]
source_pkgs = ["atlas_init"]
branch = true
parallel = true

[tool.coverage.paths]
atlas_init = ["atlas_init"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[tool.hatch.build.targets.wheel]
only-include = ["atlas_init"]

[tool.hatch.build.targets.wheel.hooks.custom]
path = "file_copy.py"
require-runtime-dependencies=true

[tool.ruff]
extend-exclude = ["test_atlas_init"]

[tool.ruff.lint]
extend-ignore = [
  "TCH003",
  "G004",
  "TRY003",
  "S101",
  "EM102",
  "FBT001",
  "B008",   # https://docs.astral.sh/ruff/rules/function-call-in-default-argument/
  "TCH002", # https://docs.astral.sh/ruff/rules/typing-only-third-party-import/
  "TCH001", # https://docs.astral.sh/ruff/rules/typing-only-first-party-import/
]
